requirements.txt
Flask==3.1.1
mysql-connector-python==9.5.0
PyYAML==6.0.2
